Made iOS app navbar and titles match web app (#14)
The navigation bar on the current version of the app is very outdated compared to the current web app. Modify the iOS navigation bar item order to match the order of items on the web application.

Another area of asymmetry is that the default page that the app opens to is to a location we no longer have a sensor in - the location still works because it pulls from a URL for a sensor that is still active, but I modified the title to match the current first index and log another bug (see Issue #15) about automating that.